move login page to component
