+++ pages/login.js
@@ -1,92 +1,6 @@
-import React, { useState } from "react"
-import { useRouter } from "next/router"
-import { connect } from "react-redux"
+import React from "react"
+import Login from "../components/Login"
 
-import { successAlert, errorAlert } from "../actions/alerts"
-// moved login to action, which calls the service
-// import { authUser } from "../services/auth"
-import { login } from "../actions/auth"
-
-function Login({
-  // state
-  alert,
-  auth,
-
-  // actions
-  errorAlert,
-  successAlert,
-
-  // auth
-  loginAction,
-}) {
-  const router = useRouter()
-
-  const [username, setUsername] = useState("")
-  const [password, setPassword] = useState("")
-
-  const handleSubmit = async (e) => {
-    e.preventDefault()
-
-    const successRedirect = () => router.push("/")
-
-    loginAction(username, password, successRedirect)
-  }
-
-  const handleChangeUsername = (e) => {
-    setUsername(e.target.value)
-  }
-
-  const handleChangePassword = (e) => {
-    setPassword(e.target.value)
-  }
-
-  return (
-    <div>
-      <h1>Login</h1>
-      <form onSubmit={handleSubmit}>
-        <div>
-          Username
-          <input type="text" onChange={handleChangeUsername} value={username} />
-        </div>
-        <div>
-          Password
-          <input
-            type="password"
-            onChange={handleChangePassword}
-            value={password}
-          />
-        </div>
-        <button type="submit">Login</button>
-      </form>
-    </div>
-  )
+export default function LoginPage() {
+  return <Login />
 }
-
-function mapStateToProps(state) {
-  console.log("state: ", state)
-
-  return {
-    alert: state.alert,
-    auth: state.auth,
-  }
-}
-
-// const actionCreators = {}
-// const mapDispatchToProps = (dispatch) => ({
-//   successAlert: (message) => dispatch(successAlert(message)),
-//   errorAlert: (message) => dispatch(errorAlert(message)),
-//   blahSample: (message) => dispatch(blahSample(message)),
-//   blahThunk: (message) => dispatch(blahThunk(message)),
-// blahSample,
-// })
-
-// optionally, mapDispatchToProps object
-const mapDispatchToProps = {
-  successAlert,
-  errorAlert,
-  loginAction: login,
-}
-
-const connectedLoginPage = connect(mapStateToProps, mapDispatchToProps)(Login)
-
-export default connectedLoginPage
